Move remaining misc tests out of runtests.c
It is unclear why these are misc tests rather than basic or
namespace tests.  They are also disappointingly cavalier about
checking return values (e.g. from the unnecessary mallocs). At
some point they need revisiting and tidying up.
