+++ expat/tests/handlers.c
@@ -253,6 +253,26 @@ start_ns_clearing_start_element(void *userData, const XML_Char *prefix,
   XML_SetStartElementHandler((XML_Parser)userData, NULL);
 }
 
+void XMLCALL
+start_element_issue_240(void *userData, const XML_Char *name,
+                        const XML_Char **atts) {
+  DataIssue240 *mydata = (DataIssue240 *)userData;
+  UNUSED_P(name);
+  UNUSED_P(atts);
+  mydata->deep++;
+}
+
+void XMLCALL
+end_element_issue_240(void *userData, const XML_Char *name) {
+  DataIssue240 *mydata = (DataIssue240 *)userData;
+
+  UNUSED_P(name);
+  mydata->deep--;
+  if (mydata->deep == 0) {
+    XML_StopParser(mydata->parser, 0);
+  }
+}
+
 /* Text encoding handlers */
 
 int XMLCALL
